Keep the listing decision out of the callable (#906)
`unit / vitest` on CI ran `functions/src/moderate.test.ts` and could not
resolve `firebase-functions/v2/https`: the root project runs `functions/src/**`
too, and CI's `npm ci` installs only the root's dependencies. It passed here
because a `functions/node_modules` from an earlier build was still on disk —
the same invisibility that cost three red commits on main in #1349.

Every other functions test the root project runs imports a leaf —
`searchWords.js`, `moderationRequest.js`, `strikes.js`, or a `kitEdited.ts`
whose only firebase import is a type. `decidedListing` now lives in one of
its own, which is where a pure policy belongs anyway.

Reproduced by moving `functions/node_modules` aside and running the root
suite: 553 files, green.
